+++ js/utils.js
@@ -6,9 +6,17 @@ const typeTranslateRus = {
   palace: 'Дворец'
 };
 
+const typeMinPrice = {
+  bungalow: 0,
+  flat: 1000,
+  hotel: 3000,
+  house: 5000,
+  palace: 10000,
+};
+
 const isResetButton = (evt) => evt.target.matches('.ad-form__reset');
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 const isClick = (evt) => evt.type === 'click';
 
-export { typeTranslateRus, isResetButton, isEscapeKey, isClick };
+export { typeTranslateRus, typeMinPrice, isResetButton, isEscapeKey, isClick };
