+++ js/validation/validation.js
@@ -1,7 +1,7 @@
 import { form } from './form.js';
-import { isResetButton } from '../utils.js';
+import { isResetButton, typeMinPrice } from '../utils.js';
 import { validateRoom, roomSelect } from './guest-rooms.js';
-import { onTypeChange, typeMinPrice, priceInput, typeSelect } from './price-ad.js';
+import { onTypeChange, priceInput, typeSelect } from './price-ad.js';
 import { onTimeChange, fieldsetTime } from './check-time.js';
 import { sendFormData } from '../api/api.js';
 import { showSuccessMessage } from './success-form-message.js';
